add ReducerDemo🐶
