+++ src/ReducerDemo.js
@@ -1,0 +1,5 @@
+import React, { useEffect, useMemo } from "react";
+
+export default () => {
+  return "ReducerDemo";
+};
